Cleaned up code
Refactored, changed some names to make them less verbose, and added some
comments to explain whats going on in some situations.
